+++ src/main/java/com/jiro4989/tkfm/data/Rectangle.java
@@ -1,42 +1,42 @@
 package com.jiro4989.tkfm.data;
 
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 
 public class Rectangle {
-  private IntegerProperty width = new SimpleIntegerProperty(0);
-  private IntegerProperty height = new SimpleIntegerProperty(0);
+  private DoubleProperty width = new SimpleDoubleProperty(0.0);
+  private DoubleProperty height = new SimpleDoubleProperty(0.0);
 
   public Rectangle() {
     this(0, 0);
   }
 
-  public Rectangle(int w, int h) {
+  public Rectangle(double w, double h) {
     setWidth(w);
     setHeight(h);
   }
 
-  public IntegerProperty widthProperty() {
+  public DoubleProperty widthProperty() {
     return width;
   }
 
-  public IntegerProperty heightProperty() {
+  public DoubleProperty heightProperty() {
     return height;
   }
 
-  public int getWidth() {
+  public double getWidth() {
     return width.get();
   }
 
-  public int getHeight() {
+  public double getHeight() {
     return height.get();
   }
 
-  public void setWidth(int w) {
+  public void setWidth(double w) {
     this.width.set(w);
   }
 
-  public void setHeight(int h) {
+  public void setHeight(double h) {
     this.height.set(h);
   }
 }
